Exporter moved to filter - to have 2 exporters - one for idea
