+++ app/controllers/admin/budget_investments_controller.rb
@@ -20,7 +20,7 @@ class Admin::BudgetInvestmentsController < Admin::BaseController
       format.html
       format.js
       format.csv do
-        send_data Budget::Investment::Exporter.new(@investments).to_csv,
+        send_data Budget::Investment::Exporters::Project.new(@investments).to_csv,
                   filename: 'budget_investments.csv'
       end
     end
